remove extra codec pipeline
