+++ src/zarr/__init__.py
@@ -37,7 +37,7 @@ from zarr.api.synchronous import (
 from zarr.core.array import Array, AsyncArray
 from zarr.core.config import config
 from zarr.core.group import AsyncGroup, Group
-from zarr.experimental.sync_codecs import SyncCodecPipeline  # noqa: F401 (registers pipeline)
+from zarr.experimental.sync_codecs import SyncCodecPipeline  # noqa: F401 (backwards compat)
 
 # in case setuptools scm screw up and find version to be 0.0.0
 assert not __version__.startswith("0.0.0")
