+++ src/zarr/abc/store.py
@@ -527,23 +527,22 @@ class Store(ABC):
     # -----------------------------------------------------------------------
     # Synchronous IO interface (opt-in)
     #
-    # These methods enable the SyncCodecPipeline to bypass the event loop
+    # These methods enable the codec pipeline to bypass the event loop
     # entirely for store IO. The default implementations raise
     # NotImplementedError; stores that wrap fundamentally synchronous
     # operations (MemoryStore, LocalStore) override them with direct
     # implementations. Remote/cloud stores (FsspecStore) leave them as-is
     # and remain async-only.
-    #
-    # See docs/design/sync-bypass.md for the full design rationale.
     # -----------------------------------------------------------------------
 
     @property
     def supports_sync(self) -> bool:
         """Whether this store has native synchronous get/set/delete methods.
 
-        When True, ``SyncCodecPipeline.read_sync`` / ``write_sync`` will call
-        ``get_sync`` / ``set_sync`` / ``delete_sync`` directly on the calling
-        thread, avoiding the event loop overhead of the async equivalents.
+        When True, the codec pipeline's ``read_sync`` / ``write_sync`` will
+        call ``get_sync`` / ``set_sync`` / ``delete_sync`` directly on the
+        calling thread, avoiding the event loop overhead of the async
+        equivalents.
 
         Subclasses that override the sync methods below should also override
         this property to return True.
@@ -558,7 +557,7 @@ class Store(ABC):
     ) -> Buffer | None:
         """Synchronous version of ``get()``.
 
-        Called by ``SyncCodecPipeline.read_sync`` to fetch chunk bytes without
+        Called by the codec pipeline's ``read_sync`` to fetch chunk bytes without
         going through the event loop. Only called when ``supports_sync`` is
         True, so the default ``NotImplementedError`` is never hit in practice.
         """
@@ -567,7 +566,7 @@ class Store(ABC):
     def set_sync(self, key: str, value: Buffer) -> None:
         """Synchronous version of ``set()``.
 
-        Called by ``SyncCodecPipeline.write_sync`` to persist encoded chunk
+        Called by the codec pipeline's ``write_sync`` to persist encoded chunk
         bytes without going through the event loop.
         """
         raise NotImplementedError
@@ -575,7 +574,7 @@ class Store(ABC):
     def delete_sync(self, key: str) -> None:
         """Synchronous version of ``delete()``.
 
-        Called by ``SyncCodecPipeline.write_sync`` when a chunk should be
+        Called by the codec pipeline's ``write_sync`` when a chunk should be
         removed (e.g. an empty chunk with ``write_empty_chunks=False``).
         """
         raise NotImplementedError
